requirements.txt
gspread==6.2.1
oauth2client==4.1.3
python-telegram-bot==20.7
flask==3.0.0
requests==2.31.0